Move unit tests to related project
