+++ server/gx-workflow-ls-format2/package.json
@@ -6,7 +6,11 @@
   "license": "MIT",
   "dependencies": {
     "@gxwf/server-common": "*",
-    "@gxwf/yaml-language-service": "*"
+    "@gxwf/yaml-language-service": "*",
+    "vscode-languageserver-textdocument": "^1.0.4",
+    "vscode-languageserver-types": "^3.16.0",
+    "vscode-languageserver": "^7.0.0",
+    "vscode-uri": "^3.0.3"
   },
   "devDependencies": {},
   "scripts": {}
