+++ server/gx-workflow-ls-native/package.json
@@ -5,7 +5,13 @@
   "author": "davelopez",
   "license": "MIT",
   "dependencies": {
-    "@gxwf/server-common": "*"
+    "@gxwf/server-common": "*",
+    "jsonc-parser": "^3.0.0",
+    "vscode-json-languageservice": "^4.2.1",
+    "vscode-languageserver-textdocument": "^1.0.4",
+    "vscode-languageserver-types": "^3.16.0",
+    "vscode-languageserver": "^7.0.0",
+    "vscode-uri": "^3.0.3"
   },
   "scripts": {
     "webpack": "webpack",
