+++ server/gx-workflow-ls-native/tests/unit/customValidationRules.test.ts
@@ -1,6 +1,6 @@
 import { createNativeWorkflowDocument } from "../testHelpers";
-import { WorkflowOutputLabelValidation } from "../../../src/common/providers/validation/WorkflowOutputLabelValidation";
-import { TestWorkflowProvider } from "../../testWorkflowProvider";
+import { WorkflowOutputLabelValidation } from "@gxwf/server-common/src/providers/validation/WorkflowOutputLabelValidation";
+import { TestWorkflowProvider } from "../testWorkflowProvider";
 
 describe("Custom Validation Rules", () => {
   describe("WorkflowOutputLabelValidation Rule", () => {
@@ -11,20 +11,20 @@ describe("Custom Validation Rules", () => {
     });
 
     it("should not provide diagnostics when there are no steps", async () => {
-      const wfDocument = createNativeWorkflowDocument(TestWorkflowProvider.nativeJson.validation.withoutSteps);
+      const wfDocument = createNativeWorkflowDocument(TestWorkflowProvider.workflows.validation.withoutSteps);
       const diagnostics = await rule.validate(wfDocument);
       expect(diagnostics).toHaveLength(0);
     });
 
     it("should not provide diagnostics when there are no workflow_outputs in the steps", async () => {
-      const wfDocument = createNativeWorkflowDocument(TestWorkflowProvider.nativeJson.validation.withThreeSteps);
+      const wfDocument = createNativeWorkflowDocument(TestWorkflowProvider.workflows.validation.withThreeSteps);
       const diagnostics = await rule.validate(wfDocument);
       expect(diagnostics).toHaveLength(0);
     });
 
     it("should not provide diagnostics when the steps contains workflow_outputs with label", async () => {
       const wfDocument = createNativeWorkflowDocument(
-        TestWorkflowProvider.nativeJson.validation.withWorkflowOutputLabels
+        TestWorkflowProvider.workflows.validation.withWorkflowOutputLabels
       );
       const diagnostics = await rule.validate(wfDocument);
       expect(diagnostics).toHaveLength(0);
@@ -32,7 +32,7 @@ describe("Custom Validation Rules", () => {
 
     it("should provide diagnostics when the steps contains workflow_outputs without label", async () => {
       const wfDocument = createNativeWorkflowDocument(
-        TestWorkflowProvider.nativeJson.validation.withoutWorkflowOutputLabels
+        TestWorkflowProvider.workflows.validation.withoutWorkflowOutputLabels
       );
       const diagnostics = await rule.validate(wfDocument);
       expect(diagnostics).toHaveLength(2);
