+++ server/gx-workflow-ls-native/tests/unit/nativeWorkflowDocument.test.ts
@@ -1,13 +1,13 @@
 import { createNativeWorkflowDocument } from "../testHelpers";
-import { TestWorkflowProvider } from "../../testWorkflowProvider";
+import { TestWorkflowProvider } from "../testWorkflowProvider";
 
 describe("NativeWorkflowDocument", () => {
   describe("getStepNodes", () => {
     it.each([
       ["", 0],
-      [TestWorkflowProvider.nativeJson.validation.withoutSteps, 0],
-      [TestWorkflowProvider.nativeJson.validation.withOneStep, 1],
-      [TestWorkflowProvider.nativeJson.validation.withThreeSteps, 3],
+      [TestWorkflowProvider.workflows.validation.withoutSteps, 0],
+      [TestWorkflowProvider.workflows.validation.withOneStep, 1],
+      [TestWorkflowProvider.workflows.validation.withThreeSteps, 3],
     ])("returns the expected number of steps", (wf_content: string, expectedNumSteps: number) => {
       const wfDocument = createNativeWorkflowDocument(wf_content);
       const stepNodes = wfDocument.getStepNodes();
