+++ server/package.json
@@ -5,21 +5,15 @@
   "version": "0.1.0",
   "author": "davelopez",
   "license": "MIT",
-  "dependencies": {
-    "jsonc-parser": "^3.0.0",
-    "vscode-languageserver": "^7.0.0",
-    "vscode-languageserver-textdocument": "^1.0.4",
-    "vscode-languageserver-types": "^3.16.0",
-    "vscode-json-languageservice": "^4.2.1",
-    "vscode-uri": "^3.0.3"
-  },
   "scripts": {
     "compile-native-server": "webpack --config ./gx-workflow-ls-native/webpack.config.js",
     "compile-format2-server": "webpack --config ./gx-workflow-ls-format2/webpack.config.js",
     "compile": "npm run compile-native-server && npm run compile-format2-server",
     "watch-native-server": "webpack --watch --progress --config ./gx-workflow-ls-native/webpack.config.js",
     "watch-format2-server": "webpack --watch --progress --config ./gx-workflow-ls-format2/webpack.config.js",
-    "watch": "concurrently --kill-others \"npm run watch-format2-server\" \"npm run watch-native-server\""
+    "watch": "concurrently --kill-others \"npm run watch-format2-server\" \"npm run watch-native-server\"",
+    "test-unit-native": "cd gx-workflow-ls-native && npm run test-unit && cd ..",
+    "test-unit": "npm run test-unit-native"
   },
   "workspaces": [
     "gx-workflow-ls-format2",
